Move BaseTestFormats to Support/
